backend/requirements.txt
fastapi==0.115.11
python-dotenv==1.0.1
uvicorn==0.34.0
pytest==8.3.5
pytest-cov==6.1.1
python-jose==3.4.0
pytest-asyncio==0.23.6
sqlalchemy==2.0.30
asyncpg==0.29.0
